backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pyjwt
bcrypt
pydantic
dotenv